LE Pairing and connection related changes
Handle repairing to an already pairing device

Add change to handle repairing to already paired device and
does not initiate smp when a smp is already in progress

CRs-fixed: 1097178

Change-Id: I277c540cfb6e9b90477b22137eeb3af0f73421d1

Allow full dev consolidation after pairing

When BR/EDR and LE RPA device records are present
in the device records and the same remote device is
paired on a new RPA, the device consolidation happens
after the IRK is received. The current consolidation
breaks out of the loop after consolidating one device
record. In such cases, BR dev record is left and when
BR link key is derived, it gets updated in a separate
device record. Since two device records are present at
the same time for the same remote, the keys and static
addresses get picked up from incorrect node and issues
such as "remove device from resolving list seen with {0}
static address" are seen.

The fix includes allowing the host to complete
the consolidation for all device records rather than
breaking at the fist match.

CRs-Fixed: 1056433
Change-Id: Ib0bf2e4029e448670ee44f63110e5332aa972580

Bluetooth: Use current address as the active address for BLE ACL

This change uses the active address, if available, once the LE
connection is completed and the local acl block is updated with
the active address. This active address is used for the pairing
mechanism and DHKey check calculations.
Issues are seen sometimes, when the new set of pairing occurs,
and finds the old device record, the node's main bda is the static
address of the remote, which gets incorrectly used for pairing.

Change-Id: If93f60454006184d45f8e58fc9c35935ae779f5e

Squash change for LE Pairing and connected related changes
Change-Id: Ib137ca15503e9b14ba74a0b9364b8467ba39c822