HOGP profile related changes
Bluetooth: Clear notification registrations on Gatt Disconnect

This patch clears all the notifications registered for a
specific gatt_if, once the gatt disconnection occurs.
In absence of this patch, the notifications registered for
the gatt interface remain "in_use" even after pairing,
which leads to failure in registration of notifications
after multiple pair/unpair attempts to HID devices

CRs-Fixed: 1011904
Change-Id: If16469c6980862b88609c6fba6638fe768ce09b7

Clear notification registrations based on BD address

Notification registrations are cleared currently for all
the connected remote LE devices when anyone of the LE
devices are disconnected. Since the registrations are
cleared, the notifications from other LE devices which are
still connected are not processed by BT stack.
This patch clears notification registrations only for the
disconnected device based on Bluetooth device address
during GATT close.

CRs-Fixed: 2038482

Change-Id: Iae04a0266ba4efbf72f1527f7c2a057faafe9564

Bluetooth: Increase profile connection timeout

This change increases the timeout after the gatt
discovery complete from 1 sec to 5 sec.
For profiles such as HOGP, the profile level gatt connections
after the discovery complete callbacks sometime take more time,
due to which the "no app timeout" disconnects the physical link.
This change will hold the gatt connection used for remote discovery,
for 5 seconds, after the discovery is complete, and thus provide
sufficient time for the profile level connections to kick in.

CRs-Fixed: 1050047
Change-Id: I8f3c3ad162e3dadeb0e3215bf9fe6a20e4d9c425

Bluetooth: Modify logic to ignore HID notifications

This change reverts the previous logic and delays the
ignoring of indications for HID devices, after the processing
of service change indications.
This is needed for keeping the update count equal to the
number of apps registered, so that service change indications
can be properly confirmed and followed by remote service
discovery.
This reverts commit a85bd580013944b321d88c3c5515a015c9f67f72.
CRs-Fixed: 982471

Change-Id: Iabba84aac1af6847bb44fd2169bd0870d8d14a55

Bluetooth: Add Dialog mouse & KB to blacklist for LE secure connections

This change allows the host to mask the Cross key bit in
the key distribution fields of the pairing request and
response while pairing with the Dialog keyboard and mouse
to prevent the remote from rejecting the DUT's pairing request.

CRs-Fixed: 1033619
Change-Id: I888b5ce8c12652ef5c393b304bed8948c4e6b444

Squash commit for the above HOGP changes
Change-Id: Ic5e56501f38f746755e9cf170d791a38fc7ea7b9