Fix service handle in service deleted callback
When a GATT server is unregistered, the services associated
with the server are deleted in BT stack.
Service handle is not added by BT stack in the service
deleted callback while sending this callback to upper
Java layers. Hence, the entry corresponding to the service
handle is not deleted in Java layer database.

The processing of subsequent GATT operations are affected
in Java layer due to incorrect database entries.
This change adds proper service handle in the service deleted
callback in BT stack.

Change-Id: I8906b302b61e99dcb15e80a9bb0b31f707350cca