A2DP: Honor remote start
- When remote initiates AVDTP start, a 3 second timer is started.
If the a2dp streaming is started within this time, then the timer
will be cancelled. Else if timer expires, AVDTP suspend will be
issued to the remote.
- Do not send offload VSC to SoC if remote started is true. Wait
till audio hal issues suspend and start for reconfig a2dp
setparam. For reconfig start request, if remote start timer is
set then cancel the timer and issue start req to send offload
VSC to SoC.
- Do not start stream request to remote started device during
handoff as it is already remote started and moved to started
state.
- Pass AV handle to a2dp on started function and use the handle
to send offload request to corresponding device index.

Change-Id: I6ab9e57ea3f27a4f0ef4fd59ba0124424bfcd2c5