Prevent deadlock between the Tx and Rx threads in HCI layer
In HCI layer, both Tx and Rx tried to acquire commands_pending_response_mutex
simultaneously. Since there may be a handshake between Tx and Rx
in the BT Transport layer, so it caused a deadlock between HCI layer's
Tx and Rx threads. Fixed it by decreasing the scope of this mutex
in Tx thread

Change-Id: I4e22fb47ea65c1c8a15e02f1d44bb0ea00cc7f37
Crs-fixed: 2103513