Disconnect LE link if cancel connect request cannot be processed
When LE connection gets established at HCI layer, the L2CAP/GATT
layers are notified about it only after read remote features
is complete. In the meantime, if there is a GATT_Disconnect request
(will result in cancel create request) that reaches the GATT/L2CAP
layer, it is not getting honored since there is no pending connection
at HCI layer.
The App thinks that the cancel create request went through fine, but
the stack keeps the connection up. In such cases, make sure that there
is a disconnection request issued to the controller.

Change-Id: I899e08e1ecda94b594156cbeaafa080bee21bfd9