Reduce delay in acking MM on stream open in non-split a2dp mode.
Reduce delay in acking MM on stream open for non-split a2dp mode.
More delay introduced in this path for AAC config will not be
required for non-split path as the same is applicable to AAC codec,
which is not supported in non-split mode.
Moreover this change will align the implementation with AOSP too.

CRs-Fixed: 2098615
Change-Id: Ia1467a0da1af897a73c00d4ea1aa3e24c412b13c