Proper update of suspend/stopped flag on codec config change
On Codec config change when same SEP is retained, Suspend is
initiated from BTA not BTIF, hence as per existing implementation
same is wrongly considered as remote initiated suspend in btif
even though it is DUT initiated one. Consideration of suspend event
as suspend initiated from remote device is to be done only when
both BTA and BTIF flag for tracking local suspend flag is not
set.

CRs-Fixed: 2155254

Change-Id: Id901f2211d04e1f99d317f95513bb6f67f193cfc