AV: Stop media task when streaming device is disconnected.
When streaming device is disconnected in dual a2dp connection
case, the media task was not stopped. Only new packets were not
enqueued and media task timer was stopped. This causes some
packets to remain in tx_audio_queue which was later attempted
to consume upon next connection. If the codec used earlier
and in current connection differs in RTP header byte, parsing
is incorrect and it may cause issues if invalid memory is attempted
to be accessd due to wrong calculations of header offset.

CRs-Fixed: 2226617
Change-Id: I9cec69c8c7264d616833e59b6d4a4f210fa32d8a