AV: Stop media task when device disconnected while suspend cfm pending.
In certain cases if device is disconnected while suspend cfm is pending
either due to DUT not able to send request or remote is not able to
respond, media task was not stopped and UIPC socket is not removed.
This causes socket not established properly during next time.
The fix handles this corner case and ensure that media task is
stopped.

CRs-Fixed: 2230100

Change-Id: Ic429189c99031d74aca88c856f53d22f71704a3a