TWS-A2DP: Skip index in closing state for connect request
Fix for a2dp connection failure when existing single connection
is overrided with new connection. A2dp SM for each device is not
synchronized and index in closing state is picked up as valid
index to send connect request which is not handled in closing
state. Skip the idex if it is in closing state.

Change-Id: Ia6ca371ca1c4b19dd63ad6990e52d3ae173ab2dd