TWS_A2DP: Set local suspend flag for TWS+ pair while pending start
During quick a2dp start and suspend, one of the earbud's start
response might be delayed. Fix make sure to set the local suspend
flag to both the TWS+ pair so that suspend will be triggered to
earbud that delayed start response once the start evt is received
from that earbud.

Change-Id: I08e2f631277edd8714e4810b44eaac638583fcdd
CRs-Fixed: 2354018