Properly initialize remote BD address if index is not valid.
When fetching btif_get_latest_playing_device() if index is not valid
need to fill the BD address to null in proper fashion, so that it can
reflect in the caller APIs. In the current way it will not reflect in
the caller APIs. So init the BDaddress in proper way when fetching
btif_get_latest_playing_device() API.

CRs-Fixed: 2366274
Change-Id: I55c5cd4fbf07aabe9caef35b604facecdc89c744