AVRCP: Fix for DUT not doing avrcp sdp discovery
Certain remote devices delay initiating avrcp connection after
a2dp connection. DUT waits for 3 sec and then initiate avrcp
sdp. If the remote has initiated avrcp connection just before sdp
is complete, DUT will not initiate connection. But disc db in use
bit is not reset. DUT cannot initiate avrcp sdp for current and
and successive connection. So peer features are not available.
This results in DUT not registering for volume change event
notification as the remote features are not available. Fix is to
reset disc db in use parameter if incoming connection is in
progress

Change-Id: Ib949df4814e542ce71366751370920f508cc6f9c