SDP: Compare continuetion offset with updated sdp records length
This change is used to fix a corner case hit with a remote device
which is requesting maximum attribute data length in sdp (depending
on its L2CAP SDP MTU)in a packet such a way that final fragment of
sdp data has length less that blacklisted length (14 bytes of
additional PBAP 1.2 record attributes length). This final smaller
fragement was not getting sent as continuetion bit was getting set
to 0.

Change-Id: I1fbb31b76d189a410ccb3db0b5ab4c7545227502
CRs-Fixed: 2428548