Increase BTM SCO control blocks to 6
When SCO gets connected and disconnected
with a connected headset, it goes into listening
mode using up one BTM SCO control block. When it
is done 3 times, all BTM SCO control blocks are
used. This leads to no SCO connection initiation with
4th connected headset. Increase BTM SCO control
blocks to 6 to be able to support 5 HFP connections.

CRs-Fixed: 2448865
Change-Id: If30589556a2fd58afcc40eb8e23304981e117124