BT: SMP_RSP_TOUT happening due to LinkKey Refresh complete event
After BREDR pairing, SMP through BREDR is trigerring. When SMP
Pairing is in process, Host receiving KEY REFRESH COMPLETE event
from the controller. Due to that, we are changing the linkkey type
to 4 from 7. Due to that in SMP, LTK derivation is failing.

Fix:
To avoid this scenario, delaying the link key type change till
SMP Pairing procedure completed. After completion, linkeytype
is changed to 4 and notifying the KEY REFRESH Notification to
upper layer.

Change-Id: I71af31911017e2a1308b006003b2d019686d9e22