A2dpSink: Trigger SHO while interim response as playing.
When streaming is happening with one device, and local plaback
is going on in other source device and got connected with the
sink device, then SHO happened, and Sink device would wait for
the actual play status change, so that current playing device
flag would set to current one and decoder update would happend,
so that audio track would have been created. But in this case
as source sent suspend immediately after start chnaged response
has been sent as pause, which is the reason, current playing
device not updated to current one, and decoder didn't updated,
which results in audio track not created, leading to no audio
although SHO happened due to new connection, in which local
streaming was already going on. After that start has been sent
by source, and interim updated as playing, but there is no trigger
left to update decoder and current playing index.

This fix would take care of triggering SHO on interim response of
playing, so that decoder update would happen and current playing
device index has been updated to current one, which results in
creation of audio track.

CRs-Fixed: 2469918
Change-Id: Ie4ce883e7be373370f87e332152643a57d5af1b6