A2dpSrc: Do endsession while BT turn off in BLE On case.
When BT turn off and BLE On, cleanup has been called, due to which
setactivedevice() API returns nothing, although it reached till stack.
Which is the reason why end_session wolud not be called where we reset
the flag is_session_started. As part of cleanup we updated end_session
to Audio HAL. As BT process would be same for the next BT turn ON,
although session has been started, BT HAL won't update to Audio HAL,
due to is_session_started flag has been set. So due to which when start
sent from DUT, it won't be delivered to remote, leading to mute issue.

This fix would take care of calling end_session when cleanup would call
in BT Turn off case. So that it would reset all the variables, so that
for next iteration onwards session start update would send to Audio HAL
and streaming would be smooth.

CRs-Fixed: 2489001
Change-Id: I850c7592b8abbcdf70707d615e15f806a215283d