Reconfig again if rcfg index was changed during reconfig
When reconfig is ongoing, if user triggers codec switch,
we will try to do another reconfig. Because the first
reconfig is ongoing, and the avdt_scb state may be at IDLE now,
the second reconfig will be ignored, but the codec we
are using has been changed to second codec because of second reconfig.
As a result, the codec set to peer is different from the codec stored
in stack. To fix it, reconfig again after reconfig is done if
rcfg index was changed.

Change-Id: I296d776dfc06c10ce5c4b84407f9ec097e805b3b
CRs-Fixed: 2506381