A2dpSrc: Send flow_spec command for the active device.
When two devices has been connected, and both are in opened state,
host is sending flow_spec command to remote on the basis of latest
connected device, irrespective of whether it is active or not.

This fix would take care, whenever setup codec has been done, host
would check for the current active device and send flow_spec command.

CRs-Fixed: 2516193
Change-Id: Ic42a12b4a8559bff179697b4f741fb7e59746758