AVRCP: Store and Update correct version based on Remote
- Earlier based on Remote feature capabilities Browsing
and Cover-Art, if none of the features are set then we
used to update remote the fallback AVRC version of 1.3
in SDP Response though storing Remote's original version
only in config file.
With this fix we will save same Remote version in conf
file and update same in SDP Response.

- Disallow absolute volume registration if stored entry
in config file in DUT is less than 1.4 for Remote.

- Disallow registering browse supported events if either
DUT's AVRC version is v1.3 OR remote's browse feature is
not supported irrespective of it's SDP version.

CRs-Fixed: 2576706
Change-Id: I07b1ef5b662e44816a0ac1c767340f592c648b57