Fix default addr type for BLE address scan filter
Issue:
When a non-paired remote device advertises with random address
and the local BLE scanner adds the remote random address as
a scan filter, adv reports are not received by the BT host from
controller.

Rootcause:
When the remote device info is not available in local BLE
scanner device database, the default addr type is set as
public in the VS command for scan filtering. Hence, the
controller does not send the remote's random address
adv report events to the host.

Fix:
When the remote device info is not available in local BLE
scanner device database, the default addr type is set as
0x02(Not applicable) in VS command for scan filtering.

Change-Id: I5f414be81bc3d12199febbc810f728b64fb0f001