A2dpSrc: Handle MM start and ALS app reconfig req gracefully.
When codec config request from App and MM start stream request
issued back to back, and when these are handled in btif, not
caching the codec config as it is just ahead of START request.
While in BTA, START reached first and sent to remote. As codec
config is little heavy process, reconfig request also sent to
remote with little delay from BTA. After reconfig request sent
to remote, remote responded for the START and in btif, AV SM
has been moved to started handler. Before remote responded for
the reconfig, MM sent start one more time due to reconfig, which
is not handled because we are already in started state in btif.
When remote responded for the reconfig with BAD_STATE, HOST will
close the AVDTP channel, and reconfig has been retried
(close->set_config->open) one more time, but there is no trigger
has been left to send START to streaming, which is the reson why
leading to no audio issue.

This issue has been fixed in below way.
- Cache codec config requested by app by-default in opened handler,
  and fake the reconfig fail to btif, so that when START VSCs has
  been completed, we can process cached codec cofig, which in-turn
  sends reconfig in proper fashion(suspend->reconfig->start)

CRs-Fixed: 2589897
Change-Id: I6c7249a4fa27f9ac61ba60586ec3742f9e8ad70f