HFP: move SCO SM to proper state if rfc is closed before disconnect SCO
If rfc is closed by remote or timeout and SCO is still open state.
SCO closed event will be ignored because peer address is empty.
As a result, SCO SM will block in SHUTTING state. To avoid this
issue, Need move SCO SM to SHUTDOWN or LISTEN state instead of
SHUTTING state if rfc is closed while SCO is open.

Change-Id: I1cd06befc812c6bb3586637139e22e011285196a
CRs-Fixed: 2606438