AVRCP: Conditional support for Player App Setting feature
If remote device version is v1.3 OR device is blacklisted
for player application settings and OR fallback on AVRCP
v1.3 due to Blacklist OR the player application settings
feature is disabled on phone, then make sure to:

- Disable Player Application settings feature bit in SDP
response of AVRCP from DUT.

- Disallow sending the PLAYER APPLICATION SETTING CHANGED
event in list of supported event for DUT in GetCapability
response.

- If despite of above two conditions met and remote still
exchanging the player application setting command, simply
send reject response.

In addition of that for Remotes having AVRCP v1.4 onwards
but browse unsupported send ADDRESSED and AVAIALBLE Player
CHANGED events to Remote in GetCapability response.

CRs-Fixed: 2606695
CRs-Fixed: 2607690
Change-Id: I8c30bfe5c340a4e5be0bde2a0e2c6c0c5960561e