A2dpSrc: Move to proper SM, while reconfig if ACL disconnection happens.
While reconfig, if there is no response from remote for AVDTP_OPEN due to remote
being in bad state and ACL is dropped, due to connection timeout STR_OPEN_FAIL
event posted to BTA, which is handled in RCFG state to re-attempt reconfig 6 more
times, and leading to reissue set_config again with different ACL handle.

Change made to fix this issue, when Reconfig fail happens, check for the ACL
connection too before retrying the reconfig. If ACL is not there move the BTA SM
to proper state.

CRs-Fixed: 2612868
Change-Id: Ie7a5004322bd10e96801563e7fbf3856177fc45a