GAP: Remove HID Mouse from database if SDP failed
Usecase: BT pairing.

Observation: SDP connection requests are collided between
remote and DUT resulting the device bond state to move to
BOND NONE. On next connection, host uses the same link key
which is used in previous connection. BOND state still be
in BONDING. That is stopping inquiry to run.

Rootcause: While SDP collision the device state is moved
to BOND NONE from BONDED, but the link is not deleted which
is resulting to use the same linkkey for new connection
which results the device to be in BONDING even when the
connection is successful.

Fix: If SDP connection request failed consequently for second
time, Remove device data base for the bonded device along with
send HCI disconnect before moving bond state to NONE.

Change-Id: Ic12fa342642b903a50aaa07fcc0ea0f5677387d4