A2DP: Don't select disabled optional codecs when local SRC reconnecting
When local SRC is establishing a connection, the BT stack follows
those default priorities to choose a codec, and A2dpService needs
to judge again after this connection up event. If a user prefers
mandatory codec, the peer state is CLOSE -> RECONFIG -> OPEN which
is redundant but lousy. Since codecs preference is saved at the
upper-layer, this change queries the mandatory codec, and raises
the priority for more preferred codec while reconnecting from local.

Change-Id: Iac6040b7327f772dfab7b3509413d9eebd7f3afd