GAP: Use local variable to update write link policy
Usecase: BT pairing.

Observations: When a remote who role switch is disabled
is paired to DUT. From this point the role switch will
be disabled for all the new connections. This results to
connection failures for the remote which expects role
switch to be enabled.

Rootcause: While sending write link policy to remote device,
it is checked whether the device is blacklisted to disable
the role switch. The same is updated to the global object
which used as a reference for other connections.

Fix: Now passing the local variable to get and update the
write link policy settings.

Change-Id: I4f59e66d0f8d7cdd523f4c4fa007ddc3f0838178