don't clear cached config if start is pending and codec changed after reconfig done
When reconfig is ongoing, second reconfig is triggered, then this second
reconfig is cached because BTIF_AV_FLAG_PENDING_START is set. After
first reconfig is done, the cached second reconfig is cleared, so
the second reconfig is ignored incorrectly.
To fix this issue, when reconfig is done, if codec_cfg_change is true and
BTIF_AV_FLAG_PENDING_START is set, we will not clear the cached
config. Once a2dp is started, btif_av_process_cached_src_codec_config
will be called to handle the cached second reconfig.

Change-Id: Iaa5752c20addcadba0d34dae4003107dc27ee4e6
CRs-Fixed: 2749146