Read maximum powers numbers from property
With this BT power values are fetched from the vendor
component via btconfigstore during BT ON. If the values
are defined in vendor component, those values are treated
as final power values. If they are not defined in vendor
component, the values fetched from the profile configuration
file will be used. If values are not defined on both, default
hardcoded values in code are used. These values are declared
in btconfigstore using a property which is protected by target
type.

Change-Id: I017635b41a03542c514336fb3b99b0d7e03d098f