Fix GATT operations execution in encryption queue
GATT operations queued in encryption queue
are not dequeued and sent to lower layers even
when LE encryption is completed.
This is because BR/EDR authentication and encryption
are in progress and the GATT operations are not
dequeued.

GATT operations are dequeued when the pending
security procedure on the corresponding
transport is completed with this fix.

Change-Id: Ib6679269a21a7f62e7d7ff02122d76be454fafaa