BLE: Send disconnection reason in the status parameter
In recent AOSP change, removed the reason code
to update in the status variable. It always send STATUS
as SUCCESS.
But due to that some adv audio unicast cases will not work
as expected. Ex: Remote goes out of range, adv audio profile
wont add the device in background whitelist.

To fix this issue, adding the previous change. i.e keeping
disconnection reason in the status variable in callback.

Change-Id: I533d8ac3b0da326efde5d6333a4ad5edb60dcf95