L2CAP : Fix for infinite loop in LE Security Access
Issue:
Issue is in pending LE security access queue, the request is dequeued
from the queue and adding it again in the same request to the same
pending queue. So it will be in loop continously for the same pending
request. This loop is not breaking in this scenario. So other tasks on
the BTU thread are not getting processed.

Fix:
Instead of adding again to the queue, will call directly request
access for security.

Change-Id: I6e95a12a93a7b4efbf55d4b5c1ac8c841110e9e4