HFP: Disconnect SCO if BTM SCO control block is unused state
In the corner scenario that DUT accept incoming SCO connection
and disconnect HFP AG connection back to back, which cause the
SCO connection complete event is not handled in btm layer and
no callback to notify bta ag layer SCO is already connected.
As as result, BTA AG SCO move to shutdown state but SCO is still
connected in lower layer.

Change-Id: Ib15cc673b05efce005080d7237dda4d8c022fc74
CRs-Fixed: 3028366